LayoutTests/accessibility/aria-current-global-attribute-expected.txt

text1 text2
This tests that aria-current is a global attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spanAccessible('text1') is false
PASS spanAccessible('text2') is true
PASS successfullyParsed is true

TEST COMPLETE
